requirements.txt
openai>=1.0.0
SpeechRecognition>=3.8.0
pyttsx3>=2.90
python-dotenv>=0.19.0
numpy>=1.21.0 